refactor: removed unnecessary component
